+++ OpenAVRc/gui/popups.h
@@ -49,9 +49,9 @@ const pm_uchar zz_asterisk_lbm[] PROGMEM = {
 #if defined(NAVIGATION_MENUS)
 const char *popupMenuItems[POPUP_MENU_MAX_LINES];
 uint8_t s_menu_item = 0;
-uint16_t popupMenuNoItems = 0;
+uint8_t popupMenuNoItems = 0;
 uint8_t popupMenuFlags = 0;
-uint16_t popupMenuOffset = 0;
+uint8_t popupMenuOffset = 0;
 #endif // defined
